moved Prose component to global & updated the article template to sue defined Prose stylings
